requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
gunicorn>=23,<24
SQLAlchemy>=2.0,<3
alembic>=1.14,<2
psycopg[binary]>=3.2,<4
python-dotenv>=1.0,<2
PyYAML>=6.0,<7
PyJWT>=2.10,<3
Werkzeug>=3.1,<4
boto3>=1.35,<2
python-multipart>=0.0.9,<1
pytest>=8.3,<9
factory_boy>=3.3,<4
httpx>=0.28,<1
